@robodoo 2025-03-14
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/201759.png)](https://mergebot.odoo.com/odoo/odoo/pull/201759)