@auma-odoo 2025-03-14
This fix is complemented by #201243