@ticodoo 2025-03-24
Hello @auma-odoo  :exploding_head: so simple, so powerful, how did we not add that already. Anyways, I'm not sure if it's worth merging this change in 16.0 since the ticket for 18.0. Maybe 17.0? (unless you have another ticket for 16.0)