@auma-odoo 2025-03-25
Hello @ticodoo :wave:. No problem, I can retarget to 17.0. It is just that since it is just a domain, it could have been good to merge it in 16.0. But I understand that you don't want to risk it.