@auma-odoo 2025-03-25
PR retargeted to 17.0 #203276